Refactor: Extract and centralize `changeNestedItems` logic for `ExpandableListItem`
- **Extracted `changeNestedItems` function**:
    - Moved the recursive logic for toggling the expanded state of an `ExpandableListItem` to a new extension function `changeNestedItems` in `ExpandableListItemExtensions.kt`.
    - This function now handles updating the `isExpanded` state and the arrow icon in the header.
- **Updated view models:**
    - Replaced the duplicate `changeNestedItems` implementation in `RequestViewModel` and `DocumentDetailsViewModel` with calls to the new centralized `changeNestedItems` extension function.
    - Removed duplicated code from `DocumentSuccessViewModel`.
    - Implemented new `changeNestedItems` extension function in ui-logic module.
