+++ common-feature/src/main/java/eu/europa/ec/commonfeature/ui/document_success/DocumentSuccessViewModel.kt
@@ -25,6 +25,7 @@ import eu.europa.ec.uilogic.component.content.ContentHeaderConfig
 import eu.europa.ec.uilogic.component.wrap.ExpandableListItem
 import eu.europa.ec.uilogic.config.ConfigNavigation
 import eu.europa.ec.uilogic.config.NavigationType
+import eu.europa.ec.uilogic.extension.changeNestedItems
 import eu.europa.ec.uilogic.mvi.MviViewModel
 import eu.europa.ec.uilogic.mvi.ViewEvent
 import eu.europa.ec.uilogic.mvi.ViewSideEffect
@@ -128,41 +129,6 @@ abstract class DocumentSuccessViewModel : MviViewModel<Event, State, Effect>() {
             copy(
                 items = updatedItems
             )
-        }
-    }
-
-    //TODO should this be in other place? i.e. should it be used elsewhere?
-    private fun List<ExpandableListItem>.changeNestedItems(id: String): List<ExpandableListItem> {
-        return this.map { nestedItem ->
-            when (nestedItem) {
-                is ExpandableListItem.NestedListItemData -> {
-                    if (nestedItem.header.itemId == id) {
-                        val newIsExpanded = !nestedItem.isExpanded
-                        val newCollapsed = nestedItem.header.copy(
-                            trailingContentData = ListItemTrailingContentData.Icon(
-                                iconData = if (newIsExpanded) {
-                                    AppIcons.KeyboardArrowUp
-                                } else {
-                                    AppIcons.KeyboardArrowDown
-                                }
-                            )
-                        )
-
-                        nestedItem.copy(
-                            header = newCollapsed,
-                            isExpanded = newIsExpanded
-                        )
-                    } else {
-                        nestedItem.copy(
-                            nestedItems = nestedItem.nestedItems.changeNestedItems(id)
-                        )
-                    }
-                }
-
-                is ExpandableListItem.SingleListItemData -> {
-                    nestedItem
-                }
-            }
         }
     }
