+++ issuance-feature/src/main/java/eu/europa/ec/issuancefeature/ui/document/details/DocumentDetailsViewModel.kt
@@ -27,11 +27,9 @@ import eu.europa.ec.issuancefeature.interactor.document.DocumentDetailsInteracto
 import eu.europa.ec.issuancefeature.interactor.document.DocumentDetailsInteractorStoreBookmarkPartialState
 import eu.europa.ec.resourceslogic.R
 import eu.europa.ec.resourceslogic.provider.ResourceProvider
-import eu.europa.ec.uilogic.component.AppIcons
-import eu.europa.ec.uilogic.component.ListItemTrailingContentData
 import eu.europa.ec.uilogic.component.content.ContentErrorConfig
 import eu.europa.ec.uilogic.component.wrap.BottomSheetTextData
-import eu.europa.ec.uilogic.component.wrap.ExpandableListItem
+import eu.europa.ec.uilogic.extension.changeNestedItems
 import eu.europa.ec.uilogic.mvi.MviViewModel
 import eu.europa.ec.uilogic.mvi.ViewEvent
 import eu.europa.ec.uilogic.mvi.ViewSideEffect
@@ -265,41 +263,6 @@ class DocumentDetailsViewModel(
                         documentClaims = updatedDocumentClaims
                     )
                 )
-            }
-        }
-    }
-
-    //TODO should this be in other place? i.e. should it be used elsewhere?
-    private fun List<ExpandableListItem>.changeNestedItems(id: String): List<ExpandableListItem> {
-        return this.map { nestedItem ->
-            when (nestedItem) {
-                is ExpandableListItem.NestedListItemData -> {
-                    if (nestedItem.header.itemId == id) {
-                        val newIsExpanded = !nestedItem.isExpanded
-                        val newCollapsed = nestedItem.header.copy(
-                            trailingContentData = ListItemTrailingContentData.Icon(
-                                iconData = if (newIsExpanded) {
-                                    AppIcons.KeyboardArrowUp
-                                } else {
-                                    AppIcons.KeyboardArrowDown
-                                }
-                            )
-                        )
-
-                        nestedItem.copy(
-                            header = newCollapsed,
-                            isExpanded = newIsExpanded
-                        )
-                    } else {
-                        nestedItem.copy(
-                            nestedItems = nestedItem.nestedItems.changeNestedItems(id)
-                        )
-                    }
-                }
-
-                is ExpandableListItem.SingleListItemData -> {
-                    nestedItem
-                }
             }
         }
     }
